requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
albumentations==0.5.2
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
EditorConfig==0.12.3
efficientnet-pytorch==0.7.0
einops==0.2.0
filelock==3.0.12
future==0.18.2
gdown==3.11.0
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsbeautifier==1.13.5
kiwisolver==1.3.1
lmdb==1.0.0
MarkupSafe==2.0.1
matplotlib==3.4.2
munch==2.5.0
networkx==2.5
numpy==1.20.1
opencv-python-headless==4.2.0.32
pandas==1.2.2
Pillow==8.2.0
prefetch-generator==1.0.1
pretrainedmodels==0.7.4
psycopg2==2.8.5
PyJWT==2.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tifffile==2021.3.16
torch==1.6.0+cpu
torchvision==0.9.1+cpu
tqdm==4.60.0
unicode==2.8
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
vietocr==0.3.5
Wand==0.6.6
whitenoise==5.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0



